refactor(server): split GitHub adapter into 4 layers
